+++ src/index.ts
@@ -1,5 +1,5 @@
 import Queue from './queue'
 
-export type { IQueue, QueueOptions, Worker } from './queue'
+export type { IQueue, QueueOptions, WorkerFunc } from './queue'
 
 export default Queue