+++ src/queue.test.ts
@@ -17,11 +17,11 @@ const q = new Queue()
 
 const Worker = {
   empty: async () => { },
-  delay: async () => { await delay(10) },
+  delay: (ms: number = 10) => async () => { await delay(ms) },
   error: async () => { throw new Error('error from worker') },
 }
 
-describe.skip('base', () => {
+describe('base', () => {
   test('should run all workers', async () => {
     const concurrency = 4
     const q = new Queue({ concurrency })
@@ -44,7 +44,7 @@ describe.skip('base', () => {
 
   test('drain should wait for all workers to finish', async () => {
     Array(n).fill(0)
-      .map(() => Worker.delay)
+      .map(() => Worker.delay(10))
       .forEach(worker => q.push(worker))
 
     // results is empty at beginning
@@ -76,7 +76,7 @@ describe.skip('base', () => {
     const q = new Queue({ autostart: true, concurrency })
 
     Array(n).fill(0)
-      .map(() => Worker.delay)
+      .map(() => Worker.delay(10))
       .forEach(worker => q.push(worker))
 
     // after push, the first `concurrency` workers is executed immediately
@@ -88,7 +88,7 @@ describe.skip('base', () => {
   })
 })
 
-describe.skip('idempotent', () => {
+describe('idempotent', () => {
   const q = new Queue({ concurrency: 1 })
 
   test('`start` should be idempotent', async () => {
@@ -126,7 +126,7 @@ describe.skip('idempotent', () => {
   })
 })
 
-describe.skip('error handling', () => {
+describe('error handling', () => {
   const n = 10
   const q = new Queue()
 
@@ -141,7 +141,7 @@ describe.skip('error handling', () => {
   })
 })
 
-describe.skip('concurrency', () => {
+describe('concurrency', () => {
   const results: number[] = []
   const workerFn = (i: number) => async () => {
     await delay((n - i) * 10)
@@ -200,6 +200,32 @@ describe('throttle', () => {
     expectAround(end(), 2000)
   })
 
+  test('trailing', async () => {
+    const q = new Queue({
+      throttler: {
+        limit: 2,
+        interval: 1000,
+      },
+    })
+    Array(5).fill(0)
+      .map((_, i) => () => delay((i % 2 === 0) ? 1000 : 500))
+      .forEach(worker => q.push(worker))
+
+    // workers should run like:
+    // 0s: 0(1.0s), 1(0.5s)
+    // 0.5s: ~1~
+    // 1.0s: ~0~
+    // 1.5s: 2(1.0s)
+    // 2.0s: 3(0.5s)
+    // 2.5s: ~2~, ~3~
+    // 3.5s: 4(1.0s)
+    // 4.5s: done
+
+    const end = timespan()
+    await q.drain()
+    expectAround(end(), 4500)
+  })
+
   test('leading', async () => {
     const q = new Queue({
       throttler: {
@@ -222,5 +248,42 @@ describe('throttle', () => {
     const end = timespan()
     await q.drain()
     expectAround(end(), 1500)
+  })
+})
+
+describe('await', () => {
+  test('base', async () => {
+    const q = new Queue()
+    expect(await q.await(async () => 1)).toBe(1)
+    expect(await q.await(async () => 'a')).toBe('a')
+  })
+
+  test('with push', async () => {
+    const q = new Queue()
+    q.push(async () => 1)
+    expect(await q.await(async () => 2)).toBe(2)
+    q.push(async () => 'a')
+    expect(await q.await(async () => 'b')).toBe('b')
+  })
+
+  test('error', async () => {
+    const q = new Queue()
+    expect(q.await(Worker.error)).rejects.toThrow('error from worker')
+  })
+
+  test('with throttler', async () => {
+    const q = new Queue({
+      throttler: {
+        limit: 1,
+        interval: 500,
+      },
+    })
+
+    q.push(Worker.empty)
+    q.push(Worker.delay(1000))
+
+    const end = timespan()
+    expect(await q.await(async () => 2)).toBe(2)
+    expectAround(end(), 500 + 1000 + 500)
   })
 })