+++ src/throttler.ts
@@ -14,26 +14,34 @@ export type ThrottlerOptions = {
   limit: number
   interval: number
   leading?: boolean
+  debug?: boolean
 }
 
 class Throttler {
   private limit: number
   private interval: number
-  private leading?: boolean
+  private leading: boolean
+  private debug: boolean
   private jobs: { [k in JobId]: Job } = {}
+  private startTime: number = Date.now()
 
-  constructor({ limit, interval, leading }: ThrottlerOptions) {
+  constructor({ limit, interval, leading = false, debug = false }: ThrottlerOptions) {
     this.limit = limit
     this.interval = interval
     this.leading = leading
+    this.debug = debug
   }
 
   private trimJobs() {
+    if (this.debug) {
+      return
+    }
+
     const now = Date.now()
     const jobIds = Object.keys(this.jobs)
     for (const id of jobIds) {
       const job = this.jobs[id]
-      if (job.end && job.start < now - this.interval) {
+      if (job.end !== null && job.end < now - this.interval) {
         delete this.jobs[id]
       }
     }
@@ -65,7 +73,7 @@ class Throttler {
       if (this.leading) {
         return now - this.interval <= job.start
       } else {
-        return now - this.interval <= job.start
+        return now - this.interval <= job.end
       }
     }).length
     return this.limit - jobNumInInterval
@@ -74,7 +82,9 @@ class Throttler {
   getNextRunTime(): number | null {
     const jobs = Object.values(this.jobs)
 
-    const runningCount = jobs.filter(job => job.end === null).length
+    const runningJobs = jobs.filter(job => job.end === null)
+    const runningCount = runningJobs.length
+    const latestRunningJobStartTime = runningCount > 0 ? runningJobs.sort((a, b) => b.start - a.start)[0].start : null
     const canRunCount = this.limit - runningCount
 
     if (this.leading) {
@@ -85,12 +95,19 @@ class Throttler {
       if (finishedJobs.length === 0) {
         return null
       }
-      return finishedJobs[Math.min(canRunCount, finishedJobs.length - 1)].end + this.interval
+      const nextRunTime = finishedJobs[Math.min(Math.max(0, canRunCount - 1), finishedJobs.length - 1)].end + this.interval
+      if (latestRunningJobStartTime !== null && nextRunTime < latestRunningJobStartTime) {
+        return null
+      }
+      return nextRunTime
     }
   }
 
   inspect() {
-    console.log(this.jobs)
+    const format = (time: number) => ((time - this.startTime) / 1000).toFixed(2) + 's'
+    console.log(Object.values(this.jobs).map(job => {
+      return `${job.id}: ${format(job.start)} - ${job.end ? format(job.end) : 'N/A'}`
+    }).join('\n'))
   }
 }